Move points to the top
