+++ h5p-editor-bildetema-words-topic-image/src/components/Ellipse/Ellipse.tsx
@@ -123,15 +123,6 @@ export const Ellipse: FC<EllipseProps> = ({
         fill="none"
         strokeWidth="0.3"
         className={styles.ellipse}
-        // onClick={event => {
-        //   if (isDrawingThisEllipse) {
-        //     return;
-        //   }
-
-        //   event.stopPropagation();
-
-        //   handleShapeClick(hotspot.word.id);
-        // }}
         onDoubleClick={()=>handleShapeClick(hotspot.word.id)}
         onMouseDown={({ clientX: x, clientY: y }) =>
           startShapeDragging(hotspot, { x, y })
