--- h5p-editor-bildetema-words-topic-image/src/components/Shape/Shape.module.scss
@@ -1,19 +1,0 @@
-.point {
-  cursor: pointer;
-  fill: orange;
-  opacity: 0.8;
-  transform-origin: 50% 50%;
-  transition: opacity 0.15s ease-in-out;
-
-  &:hover {
-    opacity: 1;
-  }
-
-  .canvas--finished & {
-    opacity: 0;
-  }
-}
-
-.ellipseStartPoint {
-  pointer-events: none;
-}
