+++ h5p-editor-bildetema-words-topic-image/src/components/Svg/Svg.module.scss
@@ -2,3 +2,23 @@
   inset: 0;
   position: absolute;
 }
+
+.point {
+  cursor: pointer;
+  fill: orange;
+  opacity: 0.8;
+  transform-origin: 50% 50%;
+  transition: opacity 0.15s ease-in-out;
+
+  &:hover {
+    opacity: 1;
+  }
+
+  .canvas--finished & {
+    opacity: 0;
+  }
+}
+
+.ellipseStartPoint {
+  pointer-events: none;
+}
